Rename stub package to federation
